add test

+++ app/controllers/invites_controller.rb
@@ -1,9 +1,9 @@
 class InvitesController < ApplicationController
   before_action :authenticate_user!
-  before_action :correct_member, only: [:new, :show]
+  before_action :correct_member, only: [:new, :create, :destroy]
   def new
     @invite = Invite.new
-    @room = Room.find(params[:id])
+    @room = Room.find(params[:room_id])
   end
   
   def create
@@ -11,12 +11,12 @@ class InvitesController < ApplicationController
     invited_user = User.find_by(unique_id: invite_params[:unique_id])
     if invited_user.nil? || invited_user.id == current_user.id
       flash[:alert] = invited_user.nil? ? 'ユーザーが存在しません、IDを確認してください' : '自分を招待することはできません'
-      redirect_to new_invite_path(room)
+      redirect_to new_room_invite_path(room)
     else
       duplicated_invite = room.invites.find_by(invited_id: invited_user.id)
       if duplicated_invite.present? || room.member?(invited_user)
         flash[:alert] = room.member?(invited_user) ? 'そのユーザーはすでに参加しています' : 'そのユーザーはすでに招待されています'
-        redirect_to new_invite_path(room)
+        redirect_to new_room_invite_path(room)
       else
         @invite = room.invites.build(
           inviting_id: invite_params[:inviting_id],
@@ -27,7 +27,7 @@ class InvitesController < ApplicationController
           redirect_to current_user
         else
           flash.now[:alert] = '招待できませんでした'
-          render new
+          redirect_to new_room_invite_path(room)
         end
       end
     end
@@ -46,11 +46,11 @@ class InvitesController < ApplicationController
   def add
     invite = Invite.find(params[:id])
     room = invite.room
-    if room.users << current_user
+    if invite.invited_user == current_user && room.users << current_user
       flash[:notice] = '参加しました'
       invite.destroy
     else
-      flash.now[:alert] = '参加できませんでした'
+      flash[:alert] = '参加できませんでした'
     end
     redirect_to current_user
   end
@@ -61,7 +61,7 @@ class InvitesController < ApplicationController
       params.require(:invite).permit(:inviting_id, :unique_id, :room_id)
     end
     def correct_member
-      room = Room.find(params[:id])
+      room = params[:room_id] ? Room.find(params[:room_id]) : Room.find(invite_params[:room_id])
       unless room.member?(current_user)
         flash[:alert] = 'その家計簿のメンバーではありません'
         redirect_to current_user
